+++ test/logic/bool3.cpp
@@ -15,7 +15,7 @@ TEST_CASE ( "bool3", "[bool3]" )
   CHECK(  b.is_true() );
   CHECK( !b.is_indeterminate() );
 
-  b = indeterminate;
+  b = indeterminate3;
   CHECK( !b.is_false() );
   CHECK( !b.is_true() );
   CHECK(  b.is_indeterminate() );
@@ -23,43 +23,43 @@ TEST_CASE ( "bool3", "[bool3]" )
 
 TEST_CASE( "Order two bool3", "[bool3]" )
 {
-  CHECK( bool3( false ) < bool3( indeterminate ) );
-  CHECK( bool3( indeterminate ) < bool3 ( true ) );
+  CHECK( bool3( false ) < bool3( indeterminate3 ) );
+  CHECK( bool3( indeterminate3 ) < bool3 ( true ) );
 }
 
 TEST_CASE( "Negate a bool3", "[bool3]" )
 {
   CHECK( !bool3( false ) == bool3( true ) );
   CHECK( !bool3( true ) == bool3( false ) );
-  CHECK( !bool3( indeterminate ) == bool3( indeterminate ) );
+  CHECK( !bool3( indeterminate3 ) == bool3( indeterminate3 ) );
 }
 
 TEST_CASE( "And of two bool3", "[bool3]" )
 {
   CHECK( bool3( true ) == ( bool3( true ) & bool3( true ) ) );
   CHECK( bool3( false ) == ( bool3( true ) & bool3( false ) ) );
-  CHECK( bool3( indeterminate ) == ( bool3( true ) & bool3( indeterminate ) ) );
+  CHECK( bool3( indeterminate3 ) == ( bool3( true ) & bool3( indeterminate3 ) ) );
 
   CHECK( bool3( false ) == ( bool3( false ) & bool3( true ) ) );
   CHECK( bool3( false ) == ( bool3( false ) & bool3( false ) ) );
-  CHECK( bool3( false ) == ( bool3( false ) & bool3( indeterminate ) ) );
+  CHECK( bool3( false ) == ( bool3( false ) & bool3( indeterminate3 ) ) );
 
-  CHECK( bool3( indeterminate ) == ( bool3( indeterminate ) & bool3( true ) ) );
-  CHECK( bool3( false ) == ( bool3( indeterminate ) & bool3( false ) ) );
-  CHECK( bool3( indeterminate ) == ( bool3( indeterminate ) & bool3( indeterminate ) ) );
+  CHECK( bool3( indeterminate3 ) == ( bool3( indeterminate3 ) & bool3( true ) ) );
+  CHECK( bool3( false ) == ( bool3( indeterminate3 ) & bool3( false ) ) );
+  CHECK( bool3( indeterminate3 ) == ( bool3( indeterminate3 ) & bool3( indeterminate3 ) ) );
 }
 
 TEST_CASE( "Or of two bool3", "[bool3]" )
 {
   CHECK( bool3( true ) == ( bool3( true ) | bool3( true ) ) );
   CHECK( bool3( true ) == ( bool3( true ) | bool3( false ) ) );
-  CHECK( bool3( true ) == ( bool3( true ) | bool3( indeterminate ) ) );
+  CHECK( bool3( true ) == ( bool3( true ) | bool3( indeterminate3 ) ) );
 
   CHECK( bool3( true ) == ( bool3( false ) | bool3( true ) ) );
   CHECK( bool3( false ) == ( bool3( false ) | bool3( false ) ) );
-  CHECK( bool3( indeterminate ) == ( bool3( false ) | bool3( indeterminate ) ) );
+  CHECK( bool3( indeterminate3 ) == ( bool3( false ) | bool3( indeterminate3 ) ) );
 
-  CHECK( bool3( true ) == ( bool3( indeterminate ) | bool3( true ) ) );
-  CHECK( bool3( indeterminate ) == ( bool3( indeterminate ) | bool3( false ) ) );
-  CHECK( bool3( indeterminate ) == ( bool3( indeterminate ) | bool3( indeterminate ) ) );
+  CHECK( bool3( true ) == ( bool3( indeterminate3 ) | bool3( true ) ) );
+  CHECK( bool3( indeterminate3 ) == ( bool3( indeterminate3 ) | bool3( false ) ) );
+  CHECK( bool3( indeterminate3 ) == ( bool3( indeterminate3 ) | bool3( indeterminate3 ) ) );
 }