+++ docs/changelog.rst
@@ -20,3 +20,4 @@ v0.1 (Not yet released)
 
 * Utils:
   - Three-valued Boolean (`bool3`)
+  - Five-valued Boolean (`bool5`)